requirements.txt
langchain
langchain-community
langchain-huggingface
pydantic
python-dotenv
streamlit
watchdog
pytest
pandas
lancedb
pyarrow
fastembed
google-generativeai
json-repair
transformers
torch
onnxruntime
accelerate